addons/work_presence_pro/requirements.txt
flask
requests
schedule
python-dateutil
gitpython
